init: add -v2onlyclearnet config option
if this option is set by the user, v1 connections to and from IPv4/IPv6
peers will be disallowed. only users with a real need should turn it on:
in the very unlikely event that everyone did, v1 nodes would struggle to
find honest peers and become easier to eclipse.