net: add option in CConman to allow only v2 clearnet connections
a boolean option `m_v2only_clearnet` is introduced in CConman
which will (in a later commit) store if the user wishes to
establish only v2 connections on IPV4 and IPV6 networks
since they are unencrypted.

this option is accessible outside CConman using
`RequiresV2Peer()` function with the network we're trying
to connect to passed as an argument. Inbound connections
need network passed as argument to detect onion inbound.